Update compile functions, and split up compile module
Compile structs (CompileBytes and Bake) are now taking mutable
references to their AST. This will probably be changed because they
don't need mutable access to ASTs.

Compile module now has bc and val, for the BC and Val structures/impls.
This is just for organizational purposes.
